Prevent undo retraction at first M101
The plugin "dimension" inserted a extrusion of filament at first M101
before any movement. This is not noticeable with low "Retraction
Distance" but with higher values a considerable amount of filament is
being extruded.
Even worse, if plugin "Alteration" inserted start gcode which moves the
head over the plate to prevent oozing during nozzle heating the melting
chamber is under pressure at start and creates a nice fat blob and/or
fat lines at start.
The retraction code should not extrude at start. This is the task of
another plugin.